clean up migrate db
